Reformatted tests with pytest
