+++ .gitignore
@@ -10,5 +10,6 @@
 
 # Local directories
 __pycache__/
+.pytest_cache/
 legacy/
 .vscode/
